Renamed default namespace
