+++ src/Outkeep.Grains.Interfaces/Governance/ActivityPriority.cs
@@ -1,4 +1,4 @@
-﻿namespace Outkeep.Grains.Governance
+﻿namespace Outkeep.Governance
 {
     public enum ActivityPriority
     {
