+++ src/Outkeep.Grains.Interfaces/Governance/IGrainControlExtension.cs
@@ -1,7 +1,7 @@
 ﻿using Orleans.Runtime;
 using System.Threading.Tasks;
 
-namespace Outkeep.Grains.Governance
+namespace Outkeep.Governance
 {
     public interface IGrainControlExtension : IGrainExtension
     {
