+++ src/Outkeep.Grains.Interfaces/OutkeepGrainFactoryExtensions.cs
@@ -1,5 +1,5 @@
-﻿using Outkeep.Grains.Caching;
-using Outkeep.Grains.Echo;
+﻿using Outkeep.Caching;
+using Outkeep.HealthChecks;
 using System;
 
 namespace Orleans
