+++ src/Outkeep.Grains/CacheGrain.cs
@@ -1,7 +1,8 @@
 ﻿using Orleans;
 using Orleans.Concurrency;
 using Orleans.Runtime;
-using Outkeep.Grains.Caching;
+using Outkeep.Caching;
+using Outkeep.Governance;
 using Outkeep.Grains.Governance;
 using System;
 using System.Threading.Tasks;
