+++ src/Outkeep.Grains/Governance/Memory/MemoryResourceGovernor.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Hosting;
 using Orleans;
 using Outkeep.Core.Governance;
+using Outkeep.Governance;
 using System;
 using System.Collections.Concurrent;
 using System.Threading;
