+++ src/Outkeep.Grains/Governance/WeakActivationState.cs
@@ -1,5 +1,6 @@
 ﻿using Orleans;
 using Orleans.Runtime;
+using Outkeep.Governance;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
