+++ test/Outkeep.Api.Http.Tests/EchoControllerV0Tests.cs
@@ -4,7 +4,7 @@ using Orleans;
 using Orleans.Runtime;
 using Outkeep.Api.Http.Controllers.V0;
 using Outkeep.Api.Http.Models.V0;
-using Outkeep.Grains.Echo;
+using Outkeep.HealthChecks;
 using System;
 using System.Threading.Tasks;
 using Xunit;
