+++ test/Outkeep.Grains.Tests/CacheGrainTests.cs
@@ -4,8 +4,8 @@ using Orleans;
 using Orleans.Concurrency;
 using Orleans.Runtime;
 using Orleans.Storage;
+using Outkeep.Caching;
 using Outkeep.Core.Caching;
-using Outkeep.Grains.Caching;
 using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
